added vim file and modified tmux file

+++ setup.bash
@@ -1,3 +1,11 @@
 #!/bin/bash
 
-ln -s ~/dotfiles/tmux.conf ~/.tmux.conf
+if [ ! -f ~/.tmux.conf ];
+then 
+  ln -s ~/dotfiles/tmux.conf ~/.tmux.conf
+fi
+
+if [ ! -f ~/.vimrc ];
+then 
+  ln -s ~/dotfiles/vimrc ~/.vimrc
+fi
